Made Changes to Device Initialization
Added some new changes to how we create the device, decoupling further
dependencies and beginning the groundwork to making some further
enhancements on the engine. Goal is to decouple more items over time,
but focus is now to get a working set of features: Input, Multiple
objects to display, and enabling these items to be more independent in
how they are created and added into the system.